+++ css/styles.css
@@ -8,7 +8,7 @@ body {
   background-color: #fefefe;
   padding: 20px;
   border: 1px solid #888;
-  width: 60%;
+  width: 80%;
   border-radius: 5px;
   position: absolute;
   display: flex;
@@ -21,12 +21,19 @@ body {
 
 .form-container input[type="text"],
 .form-container select {
-  width: 63%;
   padding: 8px;
   font-size: 14pt;
   border: 1px solid #ccc;
   border-radius: 4px;
   margin-bottom: 10px;
+}
+
+.noteName{
+  width: 20%
+}
+
+.noteContent{
+  width: 45%
 }
 
 .form-container button {
@@ -142,7 +149,7 @@ button {
 .modal {
   display: none;
   position: fixed;
-  z-index: 1;
+  z-index: 999;
   left: 0;
   top: 0;
   width: 100%;
@@ -169,18 +176,22 @@ button {
 }
 
 .column-2 {
-  width: 55%;
+  width: 10%;
 }
 
 .column-3 {
-  width: 15%;
+  width: 45%;
 }
 
 .column-4 {
-  width: 5%;
+  width: 15%;
 }
 
 .column-5 {
+  width: 5%;
+}
+
+.column-6 {
   width: 15%;
 }
