+++ js/NotesData.js
@@ -2,37 +2,44 @@ import { Categories } from "./constants.js";
 
 export const notesData = [
   {
+    name: "Call Mom",
     content: "Don't forget to call mom on 07-09-2023 and 12-09-2023.",
     category: Categories.Task,
     createdAt: "30.07.2023",
   },
   {
+    name: "Buy Groceries",
     content: "Remember to buy groceries on 28-07-2023.",
     category: Categories.Task,
     createdAt: "30.07.2023",
   },
   {
+    name: "New Project",
     content: "Had a great idea for a new project!",
     category: Categories.Idea,
     createdAt: "30.07.2023",
   },
   {
+    name: "Dentist Appointment",
     content:
       "I'm going to have a dentist appointment on 30-05-2023, I moved it from 31-07-2023.",
     category: Categories.Task,
     createdAt: "30.07.2023",
   },
   {
+    name: "Random Thought",
     content: "This is just a random thought.",
     category: Categories.Thought,
     createdAt: "30.07.2023",
   },
   {
+    name: "Project Proposal",
     content: "Need to submit the project proposal by 02-08-2023.",
     category: Categories.Task,
     createdAt: "30.07.2023",
   },
   {
+    name: "Learning Programming Language",
     content: "Start learning a new programming language.",
     category: Categories.Idea,
     createdAt: "30.07.2023",
